Add connections UI pieces
Currently not super functional, but shows all the connection categories,
and allows you to click into each category, revealing each answer one at
a time, and displays the points value.

Doesn't yet post back to the API with the categories selected, so
capability to disable the categories once selected is only partially
functional.